Fixed now() function

+++ src/util/util.c
@@ -4,11 +4,12 @@
 #include <util/util.h>
 #include <stdlib.h>
 #include <time.h>
+#include <util/logger.h>
 
 static struct timespec util__now;
 
 double util_now(void)
 {
     (void) clock_gettime(CLOCK_REALTIME, &util__now);
-    return util__now.tv_sec + util__now.tv_nsec / 1000000000.0f;
+    return ((double) util__now.tv_sec) + (util__now.tv_nsec / 1000000000.0f);
 }
